requirements.txt
playwright>=1.52.0
beautifulsoup4>=4.12.0
pymongo>=4.6.0
python-dotenv>=1.0.0
requests>=2.31.0
psycopg2-binary>=2.9.9
boto3>=1.34.0
